refactor(agents-settings): data-driven install-strategy registry
Replace triple strategy switches in the Agents settings page with a
HookInstallStrategy → ops table so a new on-disk schema is one registry entry.
